backend/backend-ml/requirements.txt
google-genai
tensorflow
keras
numpy
pillow
scipy
pandas
scikit-learn
uvicorn
fastapi
dotenv